Rename whitespace to preprocess
